5.1. Add page controller
