+++ src/main.js
@@ -1,106 +1,19 @@
-import FilmCardComponent from './components/film-card';
-import FilmPopupComponent from './components/film-popup';
-import FilmsContainerComponent from './components/films-main';
-import FilmsListComponent from './components/films-list';
-import FilmsListExtraComponent from './components/films-list-extra';
-import LoadMoreButtonComponent from './components/load-more';
-import MainNavComponent from './components/main-nav';
-import NoFilmsComponent from './components/no-films';
 import ProfileComponent from './components/profile';
-import SortComponent from './components/sort';
-
+import PageController from './controllers/page-controller';
 import {generateFilm} from './mocks/films';
-import {filters} from './mocks/filters';
-
 import {generateArray} from './utils/random';
 
 import {render} from './utils/dom';
 
 
 const CARDS_COUNT = 18;
-const CARDS_ON_START_COUNT = 5;
-const CARDS_ON_CLICK_COUNT = 5;
 
 
-const renderFilmCard = (filmContainer, film) => {
-  const escKeyHandler = (evt) => {
-    if (evt.key === `Escape` || evt.key === `Ecs`) {
-      filmPopupComponent.removeElement();
-      document.removeEventListener(`keydown`, escKeyHandler);
-    }
-  };
-
-  const filmCardComponent = new FilmCardComponent(film);
-  const filmPopupComponent = new FilmPopupComponent(film);
-
-  filmCardComponent.setClickHandler(() => {
-    render(document.body, filmPopupComponent);
-    filmPopupComponent.setCloseButtonClickHandler(() => {
-      filmPopupComponent.removeElement();
-      document.removeEventListener(`keydown`, escKeyHandler);
-    });
-    document.addEventListener(`keydown`, escKeyHandler);
-  });
-
-  render(filmContainer, filmCardComponent);
-};
-
-const renderFilmsContainer = (filmsContainer, films) => {
-  const filmsList = new FilmsListComponent();
-  render(filmsContainer, filmsList);
-
-  if (films.length === 0) {
-    render(filmsList.getElement(), new NoFilmsComponent());
-    return;
-  }
-
-  render(filmsContainer, new FilmsListExtraComponent(`Top rated`));
-  render(filmsContainer, new FilmsListExtraComponent(`Most commented`));
-
-  const loadMoreButtonComponent = new LoadMoreButtonComponent();
-  render(filmsList.getElement(), loadMoreButtonComponent);
-
-
-  const filmsListContainer = document.querySelector(`.films-list__container`);
-  let cardsShownCount = CARDS_ON_START_COUNT;
-
-  films.slice(0, cardsShownCount)
-    .forEach((film) => renderFilmCard(filmsListContainer, film));
-
-  loadMoreButtonComponent.setClickHandler(() => {
-    const prevCardsShownCount = cardsShownCount;
-    cardsShownCount += CARDS_ON_CLICK_COUNT;
-    films.slice(prevCardsShownCount, cardsShownCount)
-      .forEach((film) => renderFilmCard(filmsListContainer, film));
-
-    if (cardsShownCount >= films.length) {
-      loadMoreButtonComponent.removeElement();
-    }
-  });
-
-
-  const filmsExtraListsContainers = document.querySelectorAll(`.films-list--extra .films-list__container`);
-  const [topRatedContainer, mostCommentedContainer] = Array.from(filmsExtraListsContainers);
-
-  const filmsByRating = films.slice()
-    .sort((filmA, filmB) => filmB.rating.valueOf() - filmA.rating.valueOf());
-  renderFilmCard(topRatedContainer, filmsByRating[0]);
-  renderFilmCard(topRatedContainer, filmsByRating[1]);
-
-  const filmsByCommentsNumber = films.slice()
-    .sort((filmA, filmB) => filmB.comments.length - filmA.comments.length);
-  renderFilmCard(mostCommentedContainer, filmsByCommentsNumber[0]);
-  renderFilmCard(mostCommentedContainer, filmsByCommentsNumber[1]);
-};
-
-
-const header = document.querySelector(`.header`);
-render(header, new ProfileComponent());
-const main = document.querySelector(`.main`);
-render(main, new MainNavComponent(filters));
-render(main, new SortComponent());
-const filmsMain = new FilmsContainerComponent();
-render(main, filmsMain);
+const headerElement = document.querySelector(`.header`);
+render(headerElement, new ProfileComponent());
+const mainElement = document.querySelector(`.main`);
+const pageController = new PageController(mainElement);
 
 const films = generateArray(generateFilm, CARDS_COUNT);
-renderFilmsContainer(filmsMain.getElement(), films);
+
+pageController.render(films);
